Update monitoring file structure
