+++ pkg/monitoring/metrics/virt-controller/component_metrics.go
@@ -1,4 +1,4 @@
-package metrics
+package virt_controller
 
 import (
 	"github.com/machadovilaca/operator-observability/pkg/operatormetrics"
