+++ pkg/virt-api/BUILD.bazel
@@ -9,8 +9,8 @@ go_library(
         "//pkg/certificates/bootstrap:go_default_library",
         "//pkg/controller:go_default_library",
         "//pkg/healthz:go_default_library",
+        "//pkg/monitoring/metrics/virt-api:go_default_library",
         "//pkg/monitoring/profiler:go_default_library",
-        "//pkg/monitoring/virt-api/metrics:go_default_library",
         "//pkg/rest:go_default_library",
         "//pkg/rest/filter:go_default_library",
         "//pkg/service:go_default_library",
