+++ pkg/virt-api/webhooks/validating-webhook/admitters/vms-admitter.go
@@ -45,7 +45,7 @@ import (
 	cdiv1 "kubevirt.io/containerized-data-importer-api/pkg/apis/core/v1beta1"
 
 	"kubevirt.io/kubevirt/pkg/instancetype"
-	"kubevirt.io/kubevirt/pkg/monitoring/virt-api/metrics"
+	metrics "kubevirt.io/kubevirt/pkg/monitoring/metrics/virt-api"
 	webhookutils "kubevirt.io/kubevirt/pkg/util/webhooks"
 	"kubevirt.io/kubevirt/pkg/virt-api/webhooks"
 	virtconfig "kubevirt.io/kubevirt/pkg/virt-config"
