+++ pkg/virt-controller/watch/BUILD.bazel
@@ -24,11 +24,11 @@ go_library(
         "//pkg/healthz:go_default_library",
         "//pkg/hooks:go_default_library",
         "//pkg/instancetype:go_default_library",
+        "//pkg/monitoring/metrics/virt-controller:go_default_library",
         "//pkg/monitoring/migration:go_default_library",
         "//pkg/monitoring/migrationstats:go_default_library",
         "//pkg/monitoring/perfscale:go_default_library",
         "//pkg/monitoring/profiler:go_default_library",
-        "//pkg/monitoring/virt-controller/metrics:go_default_library",
         "//pkg/monitoring/vmistats:go_default_library",
         "//pkg/monitoring/vmstats:go_default_library",
         "//pkg/network/namescheme:go_default_library",
@@ -127,7 +127,7 @@ go_test(
         "//pkg/apimachinery/patch:go_default_library",
         "//pkg/controller:go_default_library",
         "//pkg/instancetype:go_default_library",
-        "//pkg/monitoring/virt-controller/metrics:go_default_library",
+        "//pkg/monitoring/metrics/virt-controller:go_default_library",
         "//pkg/network/sriov:go_default_library",
         "//pkg/network/vmispec:go_default_library",
         "//pkg/pointer:go_default_library",
